apps/tes/scripts/migrate_db_data/requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas